requirements.txt
PyYAML==6.0
bing_image_downloader==1.1.1
numpy==1.22.1
pandas==1.3.5
scikit-learn==1.0.2
tqdm==4.62.3
Pillow==9.0.0

face_recognition==1.3.0
mlflow==2.10.0
apache-airflow==2.8.1
fastapi==0.109.0
pydantic==2.6.0
uvicorn==0.27.0
